requirements.txt
discord.py
openai
pi